move the vcpu configuration to Vcpu
The VMM does not need to be aware of all the magic tricks needed to
configure the Vcpu. The Vcpu configuration is now moved to the Vcpu
implementation. To achieve this separation, the Vcpu is now created from
a VcpuState object which contains all the information that might be
needed from upper layers (like for example cpuid). The state should be
better extended in the future such that it can be uniformely used for
starting fresh Vcpus, as well as resuming previously created Vcpus.

With this change the VMM no longer needs to directly access the Vcpu
functions, and no Vcpu related functionality leaks into the application
code. It should be safe to assume that all interactions should happen
from now own via the Vm object.

Signed-off-by: Andreea Florescu <fandree@amazon.com>